+++ crates/tycho-execution/contracts/test/TychoRouterTestSetup.sol
@@ -29,6 +29,7 @@ import {NativeWrapExecutor} from "../src/executors/NativeWrapExecutor.sol";
 import {LiquoriceExecutor} from "../src/executors/LiquoriceExecutor.sol";
 import {AerodromeV1Executor} from "../src/executors/AerodromeV1Executor.sol";
 import {MetricExecutor} from "../src/executors/MetricExecutor.sol";
+import {RingSwapV2Executor} from "../src/executors/RingSwapV2Executor.sol";
 // Test utilities and mocks
 import "./Constants.sol";
 import "./TestUtils.sol";
@@ -127,6 +128,7 @@ contract TychoRouterTestSetup is
     FermiSwapExecutor public fermiSwapExecutor;
     MetricExecutor public metricExecutor;
     BopAMMExecutor public bopAMMExecutor;
+    RingSwapV2Executor public ringSwapV2Executor;
 
     FeeCalculator feeCalculator;
     address routerFeeReceiver;
@@ -247,8 +249,10 @@ contract TychoRouterTestSetup is
         fermiSwapExecutor = new FermiSwapExecutor(FERMI_SWAPPER);
         metricExecutor = new MetricExecutor(METRIC_ORACLE);
         bopAMMExecutor = new BopAMMExecutor(BOPAMM_SETTLEMENT);
+        ringSwapV2Executor =
+            new RingSwapV2Executor(RING_FEW_FACTORY, RING_SWAP_FACTORY);
 
-        address[] memory executors = new address[](24);
+        address[] memory executors = new address[](25);
         executors[0] = address(usv2Executor);
         executors[1] = address(usv3Executor);
         executors[2] = address(pancakev3Executor);
@@ -273,6 +277,7 @@ contract TychoRouterTestSetup is
         executors[21] = address(fermiSwapExecutor);
         executors[22] = address(metricExecutor);
         executors[23] = address(bopAMMExecutor);
+        executors[24] = address(ringSwapV2Executor);
         return executors;
     }