+++ protocols/substreams/ethereum-balancer-v2/src/abi/base_pool.rs
@@ -1,0 +1,92 @@
+const INTERNAL_ERR: &'static str = "`ethabi_derive` internal error";
+/// Contract's functions.
+#[allow(dead_code, unused_imports, unused_variables)]
+pub mod functions {
+    use super::INTERNAL_ERR;
+}
+/// Contract's events.
+#[allow(dead_code, unused_imports, unused_variables)]
+pub mod events {
+    use super::INTERNAL_ERR;
+    #[derive(Debug, Clone, PartialEq)]
+    pub struct SwapFeePercentageChanged {
+        pub swap_fee_percentage: substreams::scalar::BigInt,
+    }
+    impl SwapFeePercentageChanged {
+        const TOPIC_ID: [u8; 32] = [
+            169u8,
+            186u8,
+            63u8,
+            254u8,
+            11u8,
+            108u8,
+            54u8,
+            107u8,
+            129u8,
+            35u8,
+            44u8,
+            170u8,
+            179u8,
+            134u8,
+            5u8,
+            160u8,
+            105u8,
+            154u8,
+            213u8,
+            57u8,
+            141u8,
+            108u8,
+            206u8,
+            118u8,
+            249u8,
+            30u8,
+            232u8,
+            9u8,
+            227u8,
+            34u8,
+            218u8,
+            252u8,
+        ];
+        pub fn match_log(log: &substreams_ethereum::pb::eth::v2::Log) -> bool {
+            if log.topics.len() != 1usize {
+                return false;
+            }
+            if log.data.len() != 32usize {
+                return false;
+            }
+            return log.topics.get(0).expect("bounds already checked").as_ref()
+                == Self::TOPIC_ID;
+        }
+        pub fn decode(
+            log: &substreams_ethereum::pb::eth::v2::Log,
+        ) -> Result<Self, String> {
+            let mut values = ethabi::decode(
+                    &[ethabi::ParamType::Uint(256usize)],
+                    log.data.as_ref(),
+                )
+                .map_err(|e| format!("unable to decode log.data: {:?}", e))?;
+            values.reverse();
+            Ok(Self {
+                swap_fee_percentage: {
+                    let mut v = [0 as u8; 32];
+                    values
+                        .pop()
+                        .expect(INTERNAL_ERR)
+                        .into_uint()
+                        .expect(INTERNAL_ERR)
+                        .to_big_endian(v.as_mut_slice());
+                    substreams::scalar::BigInt::from_unsigned_bytes_be(&v)
+                },
+            })
+        }
+    }
+    impl substreams_ethereum::Event for SwapFeePercentageChanged {
+        const NAME: &'static str = "SwapFeePercentageChanged";
+        fn match_log(log: &substreams_ethereum::pb::eth::v2::Log) -> bool {
+            Self::match_log(log)
+        }
+        fn decode(log: &substreams_ethereum::pb::eth::v2::Log) -> Result<Self, String> {
+            Self::decode(log)
+        }
+    }
+}